+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 ## 0.1.0-3.dev
 
 * Added `AuthenticationNotifier`.
-* Added `QuickDelegate.addSpacing()`.
+* Added `QuickDelegate.addSpacing()`, `.buildFooter()`, `.buildLogo()`.
 * Added `quickTr()` to a `Map<String, String>` extension to pick a translation from a map of localized strings.
 * Exported `Toast`, `ToastType`, `ToastNotifer`, `Iterable` extensions, `H2`.
 * Merging the app's own localization strings into easy_localization's.
 * Made `Toast.description` optional.
+* Changed the header padding.
 
 ## 0.1.0-2.dev